+++ src/main/java/com/buuz135/simpleclaims/Main.java
@@ -66,7 +66,7 @@ public class Main extends JavaPlugin {
             WORLDS.put(event.getWorld().getName(), event.getWorld());
             this.getLogger().at(Level.INFO).log("Registered world: " + event.getWorld().getName());
 
-            if (CONFIG.get().isForceSimpleClaimsChunkWorldMap() && !event.getWorld().getWorldConfig().isDeleteOnRemove() && ClaimManager.getInstance().canClaimInDimension(event.getWorld())) {
+            if (CONFIG.get().isForceSimpleClaimsChunkWorldMap() && !event.getWorld().getWorldConfig().isDeleteOnRemove()) {
                 this.getLogger().at(Level.INFO).log("Registered map for world: " + event.getWorld().getName());
                 event.getWorld().getWorldConfig().setWorldMapProvider(new SimpleClaimsWorldMapProvider());
             } else {