+++ CHANGELOG.md
@@ -5,6 +5,7 @@
   changes it will use the config value
 * If the force simple claims map is off, now the default map will be forced
 * Added a command to add chunk to a party based on a player name /scp add-chunk-amount <player> <amount>
+* Added a dimension blacklist config option, to blacklist which dimension players can claim chunks in
 
 # 1.0.12